+++ app/libs/extractDomain.js
@@ -1,0 +1,7 @@
+const extractDomain = (fullUrl) => {
+  const matches = fullUrl.match(/^https?:\/\/([^/?#]+)(?:[/?#]|$)/i);
+  const domain = matches && matches[1];
+  return domain ? domain.replace('www.', '') : null;
+};
+
+module.exports = extractDomain;
