+++ ui/store/actions/installed.js
