+++ ui/store/components/Installed.js
