+++ ui/store/constants/algolia.js
