+++ ui/store/helpers/getAllAppPath.js
