+++ ui/store/helpers/scanInstalledAsync.js
