+++ ui/store/reducers/installed.js
