+++ ui/store/store.html
@@ -56,6 +56,6 @@
         console.log(message);
       });
     </script>
-    <script src="bundle.js"></script>
+    <script src="store.bundle.js"></script>
   </body>
 </html>
